chore: refactor dir biome
