+++ biome.json
@@ -7,12 +7,20 @@
   },
   "files": {
     "ignoreUnknown": true,
-    "includes": ["app/**", "components/**", "lib/**"]
+    "includes": [
+      "app/**",
+      "components/**",
+      "i18n/**",
+      "lib/**",
+      "messages/**",
+      "types/**"
+    ]
   },
   "formatter": {
     "enabled": true,
     "indentStyle": "space",
-    "indentWidth": 2
+    "indentWidth": 2,
+    "lineEnding": "lf"
   },
   "linter": {
     "enabled": true,
