+++ components/features/PaperDisplay.tsx
