+++ types/index.ts
@@ -1,3 +1,3 @@
-export * from './activity';
-export * from './publication';
-export * from './common';
+export * from "./activity";
+export * from "./common";
+export * from "./publication";
